feature: jump back in Study Session api

+++ src/main/java/com/mosquizto/api/dto/request/AnswerRequest.java
@@ -14,5 +14,5 @@ public class AnswerRequest implements Serializable {
     @NotBlank
     private String definition;
 
-    private Integer responseTime;
+    private Double responseTime;
 }
